+++ wies/projects/models.py
@@ -255,3 +255,16 @@ class Service(models.Model):
 
     def get_absolute_url(self):
         return reverse("service-detail", kwargs={"pk": self.pk})
+
+
+class Note(models.Model):
+    assignment = models.ForeignKey('Assignment', models.CASCADE, related_name='notes')
+    colleague = models.ForeignKey('Colleague', models.CASCADE, related_name='notes')
+    created_at = models.DateTimeField(auto_now_add=True)
+    message = models.TextField()
+
+    class Meta:
+        ordering = ['-created_at']
+
+    def __str__(self):
+        return f"Note by {self.colleague.name} on {self.assignment.name}"